Increase precision of data sharing function.

+++ libomptarget/deviceRTLs/nvptx/src/data_sharing.cu
@@ -36,9 +36,9 @@ __device__ static unsigned getMasterThreadId() {
 }
 // The lowest ID among the active threads in the warp.
 __device__ static unsigned getWarpMasterActiveThreadId() {
-  unsigned Mask = __ballot(true);
-  unsigned ShNum = 32 - (getThreadId() & DS_Max_Worker_Warp_Size_Log2_Mask);
-  unsigned Sh = Mask << ShNum;
+  unsigned long long Mask = __ballot(true);
+  unsigned long long ShNum = 32 - (getThreadId() & DS_Max_Worker_Warp_Size_Log2_Mask);
+  unsigned long long Sh = Mask << ShNum;
   return __popc(Sh);
 }
 // Return true if this is the master thread.
